Move AuthToken generator to pinot-common
